requirements.txt
# Core packages
python-docx==0.8.11
pdfplumber==0.0.10
nltk==3.8.1
regex==2023.10.3
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
jinja2==3.1.2
pydantic==2.3.0
docxtpl==0.16.7

# Utils
tqdm==4.66.1pi
